Relocated myFavoritedMovies.css and searchMovies.css
